refactor: 컨트롤러 수정 - 중복되는 컨트롤러 정리 - 스웨거 태그 및 설명추가
